refactor(ffi, jni): group suggestion exports into suggestion/ modules
funput-ffi: suggestions/suggestion_{query,store,types}.rs become
suggestion/{engine,query,store,types}.rs; the crate root re-exports the
same flat funput_suggestion_* C surface. funput-jni: the four
suggestion_* files become suggestion/{lifecycle,query,registry,store}.rs.

ABI verified unchanged: cbindgen regenerates include/funput.h with zero
diff, and the cdylib still exports all 21 Java_app_funput_* symbols.
README module maps updated.
