+++ crates/funput-ffi/src/suggestion/types.rs
