+++ crates/funput-jni/src/lib.rs
@@ -4,8 +4,5 @@ mod composition;
 mod lifecycle;
 mod registry;
 mod settings;
-mod suggestion_lifecycle;
-mod suggestion_query;
-mod suggestion_registry;
-mod suggestion_store;
+mod suggestion;
 mod support;
