+++ crates/funput-jni/src/suggestion/lifecycle.rs
@@ -4,7 +4,7 @@ use jni::EnvUnowned;
 use jni::objects::JString;
 use jni::sys::jlong;
 
-use crate::suggestion_registry;
+use super::registry;
 use crate::support::{JavaObject, neutral, safe};
 
 #[unsafe(no_mangle)]
@@ -12,7 +12,7 @@ pub extern "system" fn Java_app_funput_funput_ime_nativebridge_PersonalSuggestio
     _env: EnvUnowned<'_>,
     _this: JavaObject<'_>,
 ) -> jlong {
-    safe(0, suggestion_registry::create)
+    safe(0, registry::create)
 }
 
 #[unsafe(no_mangle)]
@@ -24,7 +24,7 @@ pub extern "system" fn Java_app_funput_funput_ime_nativebridge_PersonalSuggestio
     safe(0, || {
         let value = neutral(env.with_env(|env| path.try_to_string(env)).into_outcome());
         (!value.is_empty())
-            .then(|| suggestion_registry::open(std::path::Path::new(&value)))
+            .then(|| registry::open(std::path::Path::new(&value)))
             .flatten()
             .unwrap_or(0)
     })
@@ -36,5 +36,5 @@ pub extern "system" fn Java_app_funput_funput_ime_nativebridge_PersonalSuggestio
     _this: JavaObject<'_>,
     handle: jlong,
 ) {
-    safe((), || suggestion_registry::destroy(handle));
+    safe((), || registry::destroy(handle));
 }
