+++ crates/funput-jni/src/suggestion/registry.rs
